Create Job board
The job board project combines HTML for structuring the web pages, CSS for styling, and JavaScript for interactivity. This project includes features for posting job openings, displaying job listings, and allowing users to interact with these listings. It's a practical project for understanding web development and learning how to create dynamic and user-friendly web applications.